move search into a single file
