+++ front/lib/actions/mcp_internal_actions/rendering.ts
