+++ front/lib/actions/mcp_internal_actions/servers/include.ts
@@ -11,6 +11,7 @@ import type {
   IncludeResultResourceType,
   WarningResourceType,
 } from "@app/lib/actions/mcp_internal_actions/output_schemas";
+import { renderRelativeTimeFrameForToolOutput } from "@app/lib/actions/mcp_internal_actions/rendering";
 import {
   findTagsSchema,
   makeFindTagsDescription,
@@ -18,7 +19,6 @@ import {
 } from "@app/lib/actions/mcp_internal_actions/servers/common/find_tags_tool";
 import {
   getCoreSearchArgs,
-  renderRelativeTimeFrameForToolOutput,
   shouldAutoGenerateTags,
 } from "@app/lib/actions/mcp_internal_actions/servers/utils";
 import { withToolLogging } from "@app/lib/actions/mcp_internal_actions/wrappers";
